feat: add 배지 저장 api

+++ src/main/java/com/umc/pureum/domain/mypage/MyPageController.java
@@ -1,8 +1,8 @@
 package com.umc.pureum.domain.mypage;
 
-import com.umc.pureum.domain.mypage.dto.GetMySentencesRes;
-import com.umc.pureum.domain.mypage.dto.PostUpdateSentenceReq;
-import com.umc.pureum.domain.mypage.dto.reponse.GetProfileResponseDto;
+import com.umc.pureum.domain.mypage.dto.response.GetMySentencesRes;
+import com.umc.pureum.domain.mypage.dto.request.PostUpdateSentenceReq;
+import com.umc.pureum.domain.mypage.dto.response.GetProfileResponseDto;
 import com.umc.pureum.domain.mypage.dto.request.PatchEditProfileReq;
 import com.umc.pureum.domain.user.service.UserService;
 import com.umc.pureum.global.config.BaseException;
@@ -62,7 +62,7 @@ public class MyPageController {
     @ApiOperation("문장 수정")
     @ApiImplicitParams({
             @ApiImplicitParam(name = "Authorization", paramType = "header", value = "서비스 자체 jwt 토큰", dataTypeClass = String.class),
-            @ApiImplicitParam(name = "sentenceId", paramType = "path", value = "문장 인덱스", example = "1", dataTypeClass = Integer.class),
+            @ApiImplicitParam(name = "sentenceId", paramType = "path", value = "문장 인덱스", example = "1", dataTypeClass = Long.class),
             @ApiImplicitParam(name = "postUpdateSentenceReq", paramType = "body", value = "문장", dataTypeClass = PostUpdateSentenceReq.class)
     })
     @ResponseBody
@@ -97,7 +97,7 @@ public class MyPageController {
     @ApiOperation("문장 삭제")
     @ApiImplicitParams({
             @ApiImplicitParam(name = "Authorization", paramType = "header", value = "서비스 자체 jwt 토큰", dataTypeClass = String.class),
-            @ApiImplicitParam(name = "sentenceId", paramType = "path", value = "문장 인덱스", example = "1", dataTypeClass = Integer.class)
+            @ApiImplicitParam(name = "sentenceId", paramType = "path", value = "문장 인덱스", example = "1", dataTypeClass = Long.class)
     })
     @ResponseBody
     @PatchMapping("/sentence/{sentenceId}/delete")
